fix(hashcache): wrap TryGetChangeSignals try/catch; split SignalSets; test null-guard
- Fix 1: Wrap TryGetChangeSignals to catch exceptions from root.Resolve()
- Fix 2: Split SignalSets into separate file per global-constraint
- Fix 3: Add ShouldNotBeNull() assertion in test for clearer failures
