+++ src/Arius.Core/Shared/HashCache/FileChangeSignals.cs
@@ -3,12 +3,3 @@ namespace Arius.Core.Shared.HashCache;
 /// <summary>Cheap, platform-provided change signals for one file. See <see cref="SignalSets"/>.</summary>
 [SharedWithinAssembly]
 internal readonly record struct FileChangeSignals(long CtimeTicks, string Inode, string Dev, int SignalSet);
-
-/// <summary>Provenance tag stored on a hashcache row so signals are only compared within the same source.</summary>
-[SharedWithinAssembly]
-internal static class SignalSets
-{
-    public const int None    = 0;
-    public const int Posix   = 1;
-    public const int Windows = 2;
-}
